update RDMA experiments

+++ RDMA_PINGPONG/implementation/messages.h
@@ -6,7 +6,7 @@ const char *DEFAULT_PORT = "12345";
 const size_t BUFFER_SIZE = 10;
 const int LATENCY = 1, LIMIT = 10;// 1000000;
 const int THROUGHPUT = 0, TIMER = 10;
-int total_throughput = 0;
+double total_throughput = 0;
 uint64_t start_time, end_time;
 
 uint64_t getTimeStamp() {
